(진행중)구조화 버퍼 기본 생성시에는 중간 전달용 버퍼 생성하지 않고 바로 전달하기. 첫 생성 이후에만 가지고있을수 있도록 설정

+++ Project/Engine/extern.cpp
@@ -7,7 +7,7 @@
 
 tTransform g_transform = {};
 tGlobalValue g_GlobalVal = {};
-tSBufferInfo g_arrStructBufferInfo[eSBUFFER_SHARED_CBUFFER_IDX::END] = {};
+tSBufferInfo g_arrStructBufferInfo[(UINT)eSBUFFER_SHARED_CBUFFER_IDX::END] = {};
 Matrix g_matViewProj;
 
 const eCOMPONENT_TYPE g_RenderComIdxStart = eCOMPONENT_TYPE::eCOMPONENT_MESH_RENDER;
